refactor(shm): move pure-infra constants/helpers into core::config
Step 2.5 of SHM business-decoupling. shared_config.rs previously mixed
business state (SharedConfig capacity, ChannelToSlotIndex using
PointType) with pure-infra primitives (magic constant, default SHM
path resolution, timestamp helper). Split along that line:

  core::config   — SHARED_MAGIC, DEFAULT_SHM_PATH, default_shm_path(),
                   timestamp_ms(), parent_dir_exists(),
                   DEFAULT_SNAPSHOT_INTERVAL_SECS. No voltage-model
                   dependency.
  shared_config  — SharedConfig + ChannelToSlotIndex (business). The
                   previously-defined items are re-exported here so
                   external callers stay source-compatible.

This expands the `core/` namespace's slot/bitmap/config triple, all of
which can move to a future voltage-shm-slots crate without dragging
voltage-model or voltage-routing along.
